add hostname to table
