+++ packages/manager/src/features/LoadBalancers/LoadBalancerLanding/LoadBalancerLanding.tsx
@@ -122,6 +122,9 @@ export const LoadBalancerLanding = () => {
                 <Hidden smDown>
                   <TableCell>Ports</TableCell>
                 </Hidden>
+                <Hidden smDown>
+                  <TableCell>Hostname</TableCell>
+                </Hidden>
                 <Hidden mdDown>
                   <TableCell>Regions</TableCell>
                 </Hidden>
